Deprecate existing effect markers (#1199)
Removed the use of effect markers ("register flags") from every script, and added a mechanism for identifying effects that previously had to use markers.

Costs created by the various `Cost` functions are now tracked in a local table, and can be checked using various new utility functions.

* `EFFECT_MARKER_DRAGON_RULER` and `EFFECT_MARKER_THUNDRA` - can now be checked using `Effect.HasSelfDiscardCost`
* `EFFECT_MARKER_DETACH_XMAT` - can now be checked using `Effect.HasDetachCost`
* `EFFECT_MARKER_ALLURE_LVUP` - can now be checked using `Effect.HasSelfToGraveCost`

The other markers were not necessary, as the effects they mark could be checked in other ways.

* `EFFECT_MARKER_CARDIAN` - by checking if the effect's type is `EFFECT_SPSUMMON_PROC`
* `EFFECT_MARKER_TELLAR` - by checking if the effect's code is `EVENT_SUMMON_SUCCESS`

The old mechanism for marking effects is kept in even if unused for the time being. `Cost.Detach` is also renamed to `Cost.DetachFromSelf` and various scripts that had to be edited were modernized.